Finished: fix cannot connect to web provider bug in deployer. The bug is I need to set _url_toDeploy, not this.url_toDeploy when initializing web3 object. Next: Make sure verification is good after modularization.
