+++ tests/ustrpctest.js
@@ -2,15 +2,20 @@ var Web3=require('web3');
 
 
 
-var web3_instance=new Web3();
+// var web3_instance=new Web3(new Web3.providers.HttpProvider('http://localhost:7545'));
 
-web3_instance.setProvider(new web3_instance.providers.HttpProvider('https://fathomless-depths-36951.herokuapp.com/'));
+// console.log(web3_instance.isConnected());
+
+// web3_instance.setProvider(new web3_instance.providers.HttpProvider('https://fathomless-depths-36951.herokuapp.com/'));
 // web3_instance.setProvider(ganache.provider());
 
-var coinbase=web3_instance.eth.coinbase;
-console.log(coinbase);
-console.log(web3_instance.eth.getBalance(coinbase));
-console.log(web3_instance.eth.getBlock(0));
+var account=web3_instance.eth.accounts[0];
+console.log(account);
+
+// var coinbase=web3_instance.eth.coinbase;
+// console.log(coinbase);
+// console.log(web3_instance.eth.getBalance(coinbase));
+// console.log(web3_instance.eth.getBlock(0));
